# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Local Network Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Room {
  id: string;
  name: string;
  createdAt: string;
  peerCount: number;
}

export interface CreateRoomRequest {
  name: string;
}

export interface Message {
  id: string;
  roomId: string;
  senderId: string;
  senderName: string;
  senderIp: string;
  content: string;
  timestamp: string;
  protocol: string;
  frameType: string;
  payloadSize: number;
  direction: string;
}

export interface Peer {
  id: string;
  name: string;
  ip: string;
  connectedAt: string;
  roomId: string;
}

export type GetMessagesParams = {
  limit?: number;
};