# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Local Network Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  roomId: string;
  senderId: string;
  senderName: string;
  senderIp: string;
  content: string;
  timestamp: string;
  protocol: string;
  frameType: string;
  payloadSize: number;
  direction: string;
}