# lib/api-zod/src/generated/types/peer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * P2P Local Network Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface Peer {
  id: string;
  name: string;
  ip: string;
  connectedAt: string;
  roomId: string;
}